cli push: Use os.walk instead of glob; upload with full path
In python 2.7 the recursive argument on glob is not supported, so this
commit adds os.walk to discover all the files under a path.

Passing os.path.split(file)[1] to nc.upload stripped the path resulting
in a flattened file structure at the destination. This commit uses
os.path.relpath to find what the destination path should be relative to
dirc.

Finally, this fixes a bug where files with names like DSC1234.JPG were
not being uploaded because the check against the supExt list was case
sensitive.